Pinch.Internal.Builder

Description

This module implements a ByteString builder very similar to Builder except that it keeps track of its final serialized length. This allows it to allocate the target ByteString in one malloc and simply write the bytes to it.

Documentation

data Builder

A ByteString Builder that knows its final size.

runBuilder :: Builder -> ByteString

Build a ByteString from the given ByteString builder.

append :: Builder -> Builder -> Builder

Append two Builders into one.

int8 :: Int8 -> Builder

Serialize a single signed byte.

int16BE :: Int16 -> Builder

Serialize a signed 16-bit integer in big endian format.

int32BE :: Int32 -> Builder

Serialize a signed 32-bit integer in big endian format.

int64BE :: Int64 -> Builder

Serialize a signed 64-bit integer in big endian format.

doubleBE :: Double -> Builder

Serialize a signed 64-bit floating point number in big endian format.

byteString :: ByteString -> Builder

Inlcude the given ByteString as-is in the builder.

Note that because this operation is applied lazily, we will maintain a reference to the ByteString until the builder is executed.