Pinch.Internal.Parser

Description

Implements a basic parser for binary data. The parser does not do any extra book-keeping besides keeping track of the current position in the ByteString.

Documentation

data Parser a

A simple ByteString parser.

runParser :: Parser a -> ByteString -> Either String a

Run the parser on the given ByteString. Return either the failure message or the result.

int8 :: Parser Int8

Produces the next byte and advances the parser.

int16 :: Parser Int16

Produces a signed 16-bit integer and advances the parser.

int32 :: Parser Int32

Produces a signed 32-bit integer and advances the parser.

int64 :: Parser Int64

Produces a signed 64-bit integer and advances the parser.

double :: Parser Double

Produces a 64-bit floating point number and advances the parser.

take :: Int -> Parser ByteString

take n gets exactly n bytes or fails the parse.