Pinch.Internal.TType

Description

Defines the different types Thrift supports at the protocol level.

TType

data TType a where

Represents the type of a Thrift value.

Objects of this type are tagged with one of the TType tags, so this type also acts as a singleton on the TTypes. It allows writing code that can enforce properties about the TType of values at compile time.

Constructors

TBool :: TType TBool

TByte :: TType TByte

TDouble :: TType TDouble

TInt16 :: TType TInt16

TInt32 :: TType TInt32

TInt64 :: TType TInt64

TBinary :: TType TBinary

TStruct :: TType TStruct

TMap :: TType TMap

TSet :: TType TSet

TList :: TType TList

class Typeable a => IsTType a where

Typeclass used to map type-leve TTypes into TType objects. All TType tags are instances of this class.

Methods

ttype :: TType a

Based on the context in which this is used, it will automatically return the corresponding TType object.

data SomeTType where

Used when the TType for something is not known at compile time. Typically, this will be pattern matched inside a case statement and code that depends on the TType will be go there.

Constructors

SomeTType :: forall a. IsTType a => TType a -> SomeTType

ttypeEquality :: TType a -> TType b -> Maybe (a :~: b)

Witness the equality of two ttypes.

ttypeEqT :: forall a b. (IsTType a, IsTType b) => Maybe (a :~: b)

Witness the equality of two TTypes.

Implicit version of ttypeEquality.

Tags

data TBool

bool

data TByte

byte

data TDouble

double

data TInt16

i16

data TInt32

i32

type TEnum = TInt32

enum

data TInt64

i64

data TBinary

binary

type TText = TBinary

string

data TStruct

struct

type TUnion = TStruct

union

type TException = TStruct

exception

data TMap

map<k, v>

data TSet

set<x>

data TList

list<x>