Pinch.Internal.Value

Description

This module defines an intermediate representation for Thrift values and functions to work with the intermediate representation.

Documentation

data Value a where

Value maps directly to serialized representation of Thrift types. It contains about as much information as what gets sent over the wire. Value objects are tagged with different TType values to indicate the type of the value.

Typical usage will not involve accessing the constructors for this type. Pinchable must be used to construct Value objects or convert them back to original types.

Constructors

VBool :: !Bool -> Value TBool

VByte :: !Int8 -> Value TByte

VDouble :: !Double -> Value TDouble

VInt16 :: !Int16 -> Value TInt16

VInt32 :: !Int32 -> Value TInt32

VInt64 :: !Int64 -> Value TInt64

VBinary :: !ByteString -> Value TBinary

VStruct :: !(HashMap Int16 SomeValue) -> Value TStruct

VMap :: forall k v. (IsTType k, IsTType v) => !(HashMap (Value k) (Value v)) -> Value TMap

VSet :: forall a. IsTType a => !(HashSet (Value a)) -> Value TSet

VList :: forall a. IsTType a => !(Vector (Value a)) -> Value TList

data SomeValue where

SomeValue holds any value, regardless of type. This may be used when the type of the value is not necessarily known at compile time. Typically, this will be pattern matched on and code that depends on the value's TType will go inside the scope of the match.

Constructors

SomeValue :: IsTType a => !(Value a) -> SomeValue

castValue :: Typeable a => SomeValue -> Maybe (Value a)

Safely attempt to cast SomeValue into a Value.

valueTType :: IsTType a => Value a -> TType a

Get the TType of a Value.