Pinch.Protocol

Description

Protocols in Pinch only need to know how to serialize and deserialize Value objects. Types that want to be serialized into/from Thrift payloads define how to convert them to/from Value objects via Pinchable.

Documentation

data Protocol

Protocols define a specific way to convert values into binary and back.

Constructors

Protocol

Fields

serializeValue :: forall a. IsTType a => Value a -> (Int64, Builder)

Serializes a Value into a ByteString builder.

Returns a Builder and the total length of the serialized content.

serializeMessage :: Message -> (Int64, Builder)

Serializes a Message and its payload into a ByteString builder.

Returns a Builder and the total length of the serialized content.

deserializeValue :: forall a. IsTType a => ByteString -> Either String (Value a)

Reads a Value from a ByteString.

deserializeMessage :: ByteString -> Either String Message

Reads a Message and its payload from a ByteString.